form edit component looks pretty good
