+++ components/community-groups/CommunityGroupsCard.jsx
@@ -1,19 +1,15 @@
-import * as React from 'react';
-import PropTypes from 'prop-types';
+import React, { useState, useEffect } from 'react';
 import { makeStyles } from '@material-ui/core/styles';
 import {
-  Card, CardContent, Typography, Button, InputLabel, 
-  TextField, FormControl
+  Card, CardContent, Typography, Button
 } from '@material-ui/core';
 import Dialog from '@material-ui/core/Dialog';
 import DialogActions from '@material-ui/core/DialogActions';
 import DialogContent from '@material-ui/core/DialogContent';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import DialogTitle from '@material-ui/core/DialogTitle';
-import NativeSelect from '@material-ui/core/NativeSelect';
-// EditableCG here
-// import EditableEvent from '../editableEvent/EditableEvent';
-import { generateOptions } from '../Helpers';
+
+import Form from './CommunityGroupsForm';
 
 const useStyles = makeStyles({
   root: {
@@ -54,91 +50,41 @@ const generateStringOfGroup = (group) => {
   return str;
 }
 
-const generateOptionsByNames = (users) => {
-  let l = [];
-  for (let i = 0; i < users.length; i++) {
-    let name = users[i].name.first + " " + users[i].name.last;
-    l.push(<option value={name}>{name}</option>);
-  }
-  return l;
-}
-
 export default function CommunityGroupsCard(props) {
   const classes = useStyles();
 
   const [openEdit, setOpenEdit] = React.useState(false);
   const [openRem, setOpenRem] = React.useState(false);
-  // const [selectedValue, setSelectedValue] = React.useState(emails[1]);
   const { cg, users } = props;
 
-  const [cgFinal, setCgFinal] = React.useState({
-    year: cg.year,
-    leadersNames: cg.leadersNames,
-    gender: cg.gender,
-    day: cg.day,
-    dorm: cg.dorm
-  });
-
-  const leaderOptions = generateOptionsByNames(users);
-  const years = ["Freshman", "Sophomore", "Junior", "Senior"];
-  const yearOptions = generateOptions(years);
-  const gender = ["Male", "Female"];
-  const genderOptions = generateOptions(gender);
-  console.log("Year options are: ", yearOptions);
+  // const [cgFinal, setCgFinal] = React.useState({
+  //   year: cg.year,
+  //   leadersNames: cg.leadersNames,
+  //   gender: cg.gender,
+  //   day: cg.day,
+  //   dorm: cg.dorm
+  // });
 
   cg.leadersNamesString = generateStringOfGroup(cg.leadersNames);
 
   const handleClickOpenEdit = () => {
+    console.log("Edit clicked");
     setOpenEdit(true);
   };
 
-  const handleCloseEdit = () => {
+  const handleClickOpenRem = (remove) => {
+    setOpenRem(remove);
+  };
+
+  const closeForm = () => {
+    console.log("Form is closed");
     setOpenEdit(false);
-  };
-
-  const handleClickOpenRem = () => {
-    setOpenRem(true);
-  };
-
-  const handleCloseRem = (remove) => {
-    console.log("Remove CG ", remove);
-    setOpenRem(false);
-  };
-
-  const indexSelectChange = (event, index) => {
-    const value = event.target.value;
-    var cgLeaders = cgFinal.leadersNames;
-    cgLeaders[index] = value;
-    setCgFinal({
-      ...cgFinal, // necessary merging existing state with new state
-      leadersNames: cgLeaders
-    });
   }
 
-  const selectChange = (event) => {
-    const name = event.target.name;
-    const value = event.target.value;
-    // console.log("Name change is: ", name);
-    // console.log("Value change is: ", value);
-    setCgFinal({
-      ...cgFinal, // necessary merging existing state with new state
-      [name]: value
-    });
-  }
-
-  const leaderSelects = cg.leadersNames.map((name, index) => {
-    return (<div>
-      <InputLabel>Leaders</InputLabel>
-      <NativeSelect
-        value={name}
-        onChange={(event) => indexSelectChange(event, index)}
-        name={index}
-      >
-      {leaderOptions}
-      </NativeSelect>
-    </div> 
-    );
-  });
+  // const handleCloseRem = (remove) => {
+  //   // console.log("Remove CG ", remove);
+  //   setOpenRem(false);
+  // };
   
   // default to filtered to movement CRU
   return (
@@ -165,71 +111,14 @@ export default function CommunityGroupsCard(props) {
           <Button variant="outlined" color="primary" onClick={handleClickOpenEdit}>
             Edit CG
           </Button>
-          <Button variant="outlined" color="secondary" onClick={handleClickOpenRem}>
+          <Button variant="outlined" color="secondary" onClick={() => handleClickOpenRem(true)}>
             Remove CG
           </Button>
         </div>
       </Card>
-      <Dialog open={openEdit} onClose={handleCloseEdit} aria-labelledby="form-dialog-title">
-        <DialogTitle id="form-dialog-title">Edit CG</DialogTitle>
-        <DialogContent>
-          <div>
-            <FormControl className={classes.formControl}>
-              <InputLabel>Year</InputLabel>
-              <NativeSelect
-                value={cgFinal.year}
-                onChange={selectChange}
-                name='year'
-              >
-                {yearOptions}
-              </NativeSelect>
-            </FormControl>
-          </div>
-          
-          <br/>
-          <div>
-            <FormControl className={classes.formControl}>
-              <InputLabel>Gender</InputLabel>
-              <NativeSelect
-                value={cgFinal.gender}
-                onChange={selectChange}
-                name='gender'
-              >
-                {genderOptions}
-              </NativeSelect>
-            </FormControl>
-          </div>
-          
-          <br/>
-          <FormControl className={classes.formControl}>
-            {leaderSelects}
-          </FormControl>
-          <br/>
-          <TextField
-              autoFocus
-              margin="dense"
-              id="name"
-              label="Meets at"
-              type="email"
-              value={cgFinal.dorm}
-              name='dorm'
-              onChange={selectChange}
-              fullWidth
-            />
-          <br/>
-        </DialogContent>
-          
-        <DialogActions>
-          <Button onClick={handleCloseEdit} color="primary">
-            Cancel
-          </Button>
-          <Button onClick={handleCloseEdit} color="primary">
-            Update
-          </Button>
-        </DialogActions>
-      </Dialog>
-
-      <Dialog open={openRem} onClose={() => handleCloseRem(false)} aria-labelledby="form-dialog-title">
+      <Form open={openEdit} cg={cg} users={users} closeForm={closeForm}>
+      </Form>
+      <Dialog open={openRem} onClose={() => handleClickOpenRem(false)} aria-labelledby="form-dialog-title">
         <DialogTitle id="form-dialog-title">Remove CG</DialogTitle>
         <DialogContent>
           <DialogContentText>
@@ -240,20 +129,14 @@ export default function CommunityGroupsCard(props) {
 
         </DialogContent>
         <DialogActions>
-          <Button onClick={() => handleCloseRem(true)} color="primary">
+          <Button onClick={() => handleClickOpenRem(true)} color="primary">
             Yes
           </Button>
-          <Button onClick={() => handleCloseRem(false)} color="primary">
+          <Button onClick={() => handleClickOpenRem(false)} color="primary">
             No
           </Button>
         </DialogActions>
       </Dialog>
-
     </div>
   );
 }
-
-// TODO: possibly add more specific validation functions
-// CommunityGroupsCard.propTypes = {
-//   cg: PropTypes.instanceOf(CommunityGroupModel).isRequired,
-// };
